second commit after gird and docker chnages done

+++ src/test/java/testBase/BaseClass.java
@@ -3,6 +3,7 @@ package testBase;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
+import java.net.URL;
 import java.text.SimpleDateFormat;
 import java.time.Duration;
 import java.util.Date;
@@ -12,11 +13,14 @@ import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.OutputType;
+import org.openqa.selenium.Platform;
 import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.openqa.selenium.remote.RemoteWebDriver;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Parameters;
@@ -36,27 +40,62 @@ public class BaseClass {
 		FileInputStream file = new FileInputStream("./src//test//resources//config.properties");
 		p=new Properties();
 		p.load(file);
-		
 		logger=LogManager.getLogger(this.getClass());
 		
-		switch(browser.toLowerCase())
+		if (p.getProperty("execution_env").equalsIgnoreCase("remote"))
 		{
-		case "chrome":driver = new ChromeDriver();
-		break;
+		 DesiredCapabilities capabilities = new DesiredCapabilities();
+		 
+			 //operating system selection
+			 if(os.equalsIgnoreCase("windows"))
+			 {
+				 capabilities.setPlatform(Platform.WIN11);
+			 }
+			 else if(os.equalsIgnoreCase("mac"))
+			 {
+				 capabilities.setPlatform(Platform.MAC);
+			 }
+			 else if(os.equalsIgnoreCase("linux"))
+			 {
+				 capabilities.setPlatform(Platform.LINUX);
+			 }
+			 else
+			 {
+				 System.out.println("no matching os");
+				 return;
+			 }
+		  //browser
+			 switch (browser.toLowerCase())
+			 {
+			 case "chrome":capabilities.setBrowserName("chrome"); break;
+			 case "edge": capabilities.setBrowserName("MicrosoftEdge");break;
+			 case "firefox": capabilities.setBrowserName("firefox");break;
+			 default: System.out.println("no matching browser");return;
+			 }
+			 
+			 driver = new RemoteWebDriver(new URL("http://localhost:4444/wd/hub"),capabilities);
+		}
 		
-		case "edge": driver = new EdgeDriver();
-		break;
-		
-		case "firefox": driver = new FirefoxDriver();
-		break;
-		
-		default: System.out.println("Invalid browser name");
-		return;
+		if(p.getProperty("execution_env").equals("local"))
+		{
+			switch(browser.toLowerCase())
+			{
+			case "chrome":driver = new ChromeDriver();
+			break;
+			
+			case "edge": driver = new EdgeDriver();
+			break;
+			
+			case "firefox": driver = new FirefoxDriver();
+			break;
+			
+			default: System.out.println("Invalid browser name");
+			return;
+		}
 		}
 		
 		
 		driver.manage().deleteAllCookies();
-		
 		driver.get(p.getProperty("appURL"));
 		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(20));
 		driver.manage().window().maximize();
